Feature changes after first implementation (#2)
* feat: removed enable and disable script

Not used in Implementation because of SSOkey. Moved to Create and Delete script

* feat: Added delete script to disable user

* Feat: Added naming function for fullname

* Feat: changed default active included remark

* Feat: Added SsoAccountKey and changed mail to UPN

* Feat: Comment out division and aliases

division is tested at customer but not used
aliases is not used and not tested

* Feat: removed all aliases code and function

* Feat: added .Resources at the responeuser

This is done to make all scripts more compatible. The get calls with userID are returned without the resources object

* Feat: enriching response user

Instead of building our own JSON 'PUT" we now enrich the response from Zivver. Because we do this we don't compare the account object with the response anymore (only in update).

* Feat: added exportData with aRef

* Fix: alt shift F + version number

* Fix: removed Cmdletbinding

Based on feedback from connector team (new best practices)

* Feat: added naming function for fullname

* Feat: make update account the same as create

* Feat: also removed aliases from update script

* feat: removed mandatory check

Because we now enrich the response of Zivver this field is not Mandatory. If you not wish to change it in the update script.

* Fix: Correct solution to detect user not found

* Feat: Only enrich zivver response

schema's in account object are not needed anymore because they are always returned by the zivver response
Exclude properties are extended with the properties used on the implementation

* fix: alt shift f + version number

* feat: removed cmdletbinding

* Feat: removed cmdletbinding + version

* Fix: account.username m check not needed

No account object in grant script

* Fix: response on no user found now correct

* Feat: group is now enriched instead of added

Because Zivver op "delete" deletes alle members. We changed the grant and revoke script to replace members.
For grant this means get current members and add the $aref

* Fix: alt shift f + version

* Feat: removed cmdletbinding

* Fix: no account object in revoke script

* Fix: no user found is now correctly handled

* Feat: now correcly remove one user from a group

Only addition to grant script is when the latest user is removed from a group "replace" is not supported. We use a "remove" then.

* Fix: alt shift f + version

* feat: removed cmdletbinding

* Feat: updated readme after first implementation

* Fix: small fix to warning SsoAccountKey

* Fix: replaced disable to delete